requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
fastapi==0.115.5
filelock==3.16.1
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
packaging==23.2
pillow==11.0.0
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.11.6
repath==0.9.0
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
text-unidecode==1.3
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.0.0
websockets==14.1
